chore(auth): Fix/clean/rebase pr for ldap and modularizing the auth module more
