+++ lib/web/app.js
@@ -150,7 +150,7 @@ exports.registerAPI = function(dreadnot, authdb) {
 
 exports.run = function(port, dreadnot) {
   var app = express.createServer(),
-      authdb = dreadnot.config.login_required ? auth.loadDBFromFile(dreadnot.config.htpasswd_file) : auth.createProxyPassAuth(),
+      authdb = auth.initialize(dreadnot.config),
       sessionStore = new connect.session.MemoryStore({maxAge: 7 * 24 * 60 * 60 * 1000});
 
   app.configure(function() {
