+++ lib/web/auth/passthru.js
@@ -1,0 +1,32 @@
+/*
+ *  Copyright 2011 Rackspace
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+function PassThruAuthDb() {}
+
+PassThruAuthDb.prototype.validate = function(xForwardedUser, password, callback) {
+    if (xForwardedUser) {
+        callback(null, true);
+    } else {
+        callback(null, false);
+    }
+};
+
+function createProxyPassAuth() {
+    return new PassThruAuthDb();
+}
+
+exports.initialize = createProxyPassAuth;
